feat: Filter the measurement errors, save the processed data in HDFStore file and possibility to select best electrodes
Add several new features to the program: when all the channel have a value of
0.0, the previous data is taken.
When the data is processed, as it can take a lot of time for huge directories,
the data is save in a HDFStore file (binary file), which is very fast.
Electrodes can then be selected according to their individual score.
For now it seems this does not inprove the total score. However it could be
useful in later improvements.